# core-lending-credit-scoring-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-scoring
  description: Credit Scoring Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-score
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/scoring-cases/{scoringCaseId}/requests/{scoringRequestId}/results/{scoringResultId}:
    get:
      tags:
        - ScoringResult
      summary: Get a scoring result by ID
      operationId: getById_3
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringResultId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringResultDTO'
    put:
      tags:
        - ScoringResult
      summary: Update a scoring result
      operationId: update_3
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringResultId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringResultDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringResultDTO'
    delete:
      tags:
        - ScoringResult
      summary: Delete a scoring result
      operationId: delete_3
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringResultId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/scoring-cases/{scoringCaseId}/requests/{scoringRequestId}/results:
    get:
      tags:
        - ScoringResult
      summary: List or search scoring results for a request
      operationId: findAll_3
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.scoringRequestId
          in: query
          description: Exact filter for scoring request id
          required: false
          schema:
            type: string
        - name: filters.scoreValue
          in: query
          description: Exact filter for score value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[scoreValue].from
          in: query
          description: Filter score value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[scoreValue].to
          in: query
          description: Filter score value to value
          required: false
          schema:
            type: string
        - name: filters.scoreOutcome
          in: query
          description: Exact filter for score outcome
          required: false
          schema:
            type: string
            enum:
              - APPROVE
              - REJECT
              - MANUAL_REVIEW
              - REFER_TO_SUPERVISOR
        - name: filters.reasonCodes
          in: query
          description: Exact filter for reason codes
          required: false
          schema:
            type: string
        - name: filters.extraDetails
          in: query
          description: Exact filter for extra details
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseScoringResultDTO'
    post:
      tags:
        - ScoringResult
      summary: Create a new scoring result
      operationId: create_3
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringResultDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringResultDTO'
...
components:
  schemas:
...
    ScoringResultDTO:
      required:
        - scoreOutcome
        - scoreValue
        - scoringRequestId
      type: object
      properties:
        scoringResultId:
          type: string
          format: uuid
          readOnly: true
        scoringRequestId:
          type: string
          format: uuid
        scoreValue:
          maximum: 1
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        scoreOutcome:
          type: string
          enum:
            - APPROVE
            - REJECT
            - MANUAL_REVIEW
            - REFER_TO_SUPERVISOR
        reasonCodes:
          maxLength: 2000
          minLength: 0
          type: string
        extraDetails:
          maxLength: 5000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseScoringResultDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ScoringResultDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...